Move error catch and tracing to its own file
